Moved asset import views inside AssetAdmin class.
