+++ assets/models.py
@@ -219,63 +219,63 @@ class Supplier(Audit):
 
 class Model(Audit):
     type_choices = (
-            ('Air conditioner', 'Air conditioner'),
-            ('Camera - Compact', 'Camera - Compact'),
-            ('Camera - SLR', 'Camera - SLR'),
-            ('Camera - Security (IP)', 'Camera - Security (IP)'),
-            ('Camera - Security (non-IP)', 'Camera - Security (non-IP)'),
-            ('Camera - Other', 'Camera - Other'),
-            ('Chassis', 'Chassis'),
-            ('Computer - Desktop', 'Computer - Desktop'),
-            ('Computer - Docking station', 'Computer - Docking station'),
-            ('Computer - Input device', 'Computer - Input device'),
-            ('Computer - Laptop', 'Computer - Laptop'),
-            ('Computer - Misc Accessory', 'Computer - Misc Accessory'),
-            ('Computer - Monitor', 'Computer - Monitor'),
-            ('Computer - Tablet PC', 'Computer - Tablet PC'),
-            ('Computer - Other', 'Computer - Other'),
-            ('Environmental monitor', 'Environmental monitor'),
-            ('Network - Hub', 'Network - Hub'),
-            ('Network - Media converter', 'Network - Media converter'),
-            ('Network - Modem', 'Network - Modem'),
-            ('Network - Module or card', 'Network - Module or card'),
-            ('Network - Power injector', 'Network - Power injector'),
-            ('Network - Router', 'Network - Router'),
-            ('Network - Switch (Ethernet)', 'Network - Switch (Ethernet)'),
-            ('Network - Switch (FC)', 'Network - Switch (FC)'),
-            ('Network - Wireless AP', 'Network - Wireless AP'),
-            ('Network - Wireless bridge', 'Network - Wireless bridge'),
-            ('Network - Wireless controller', 'Network - Wireless controller'),
-            ('Network - Other', 'Network - Other'),
-            ('Phone - Conference', 'Phone - Conference'),
-            ('Phone - Desk', 'Phone - Desk'),
-            ('Phone - Gateway', 'Phone - Gateway'),
-            ('Phone - Mobile', 'Phone - Mobile'),
-            ('Phone - Wireless or portable', 'Phone - Wireless or portable'),
-            ('Phone - Other', 'Phone - Other'),
-            ('Power Distribution Unit', 'Power Distribution Unit'),
-            ('Printer - Fax machine', 'Printer - Fax machine'),
-            ('Printer - Local', 'Printer - Local'),
-            ('Printer - Local Multifunction', 'Printer - Local Multifunction'),
-            ('Printer - Multifunction copier', 'Printer - Multifunction copier'),
-            ('Printer - Plotter', 'Printer - Plotter'),
-            ('Printer - Workgroup', 'Printer - Workgroup'),
-            ('Printer - Other', 'Printer - Other'),
-            ('Projector', 'Projector'),
-            ('Rack', 'Rack'),
-            ('Server - Blade', 'Server - Blade'),
-            ('Server - Rackmount', 'Server - Rackmount'),
-            ('Server - Tower', 'Server - Tower'),
-            ('Storage - Disk array', 'Storage - Disk array'),
-            ('Storage - NAS', 'Storage - NAS'),
-            ('Storage - SAN', 'Storage - SAN'),
-            ('Storage - Other', 'Storage - Other'),
-            ('Speaker', 'Speaker'),
-            ('Tablet', 'Tablet'),
-            ('Tape autoloader', 'Tape autoloader'),
-            ('Tape drive', 'Tape drive'),
-            ('UPS', 'UPS'),
-            ('Other', 'Other'),
+        ('Air conditioner', 'Air conditioner'),
+        ('Camera - Compact', 'Camera - Compact'),
+        ('Camera - SLR', 'Camera - SLR'),
+        ('Camera - Security (IP)', 'Camera - Security (IP)'),
+        ('Camera - Security (non-IP)', 'Camera - Security (non-IP)'),
+        ('Camera - Other', 'Camera - Other'),
+        ('Chassis', 'Chassis'),
+        ('Computer - Desktop', 'Computer - Desktop'),
+        ('Computer - Docking station', 'Computer - Docking station'),
+        ('Computer - Input device', 'Computer - Input device'),
+        ('Computer - Laptop', 'Computer - Laptop'),
+        ('Computer - Misc Accessory', 'Computer - Misc Accessory'),
+        ('Computer - Monitor', 'Computer - Monitor'),
+        ('Computer - Tablet PC', 'Computer - Tablet PC'),
+        ('Computer - Other', 'Computer - Other'),
+        ('Environmental monitor', 'Environmental monitor'),
+        ('Network - Hub', 'Network - Hub'),
+        ('Network - Media converter', 'Network - Media converter'),
+        ('Network - Modem', 'Network - Modem'),
+        ('Network - Module or card', 'Network - Module or card'),
+        ('Network - Power injector', 'Network - Power injector'),
+        ('Network - Router', 'Network - Router'),
+        ('Network - Switch (Ethernet)', 'Network - Switch (Ethernet)'),
+        ('Network - Switch (FC)', 'Network - Switch (FC)'),
+        ('Network - Wireless AP', 'Network - Wireless AP'),
+        ('Network - Wireless bridge', 'Network - Wireless bridge'),
+        ('Network - Wireless controller', 'Network - Wireless controller'),
+        ('Network - Other', 'Network - Other'),
+        ('Phone - Conference', 'Phone - Conference'),
+        ('Phone - Desk', 'Phone - Desk'),
+        ('Phone - Gateway', 'Phone - Gateway'),
+        ('Phone - Mobile', 'Phone - Mobile'),
+        ('Phone - Wireless or portable', 'Phone - Wireless or portable'),
+        ('Phone - Other', 'Phone - Other'),
+        ('Power Distribution Unit', 'Power Distribution Unit'),
+        ('Printer - Fax machine', 'Printer - Fax machine'),
+        ('Printer - Local', 'Printer - Local'),
+        ('Printer - Local Multifunction', 'Printer - Local Multifunction'),
+        ('Printer - Multifunction copier', 'Printer - Multifunction copier'),
+        ('Printer - Plotter', 'Printer - Plotter'),
+        ('Printer - Workgroup', 'Printer - Workgroup'),
+        ('Printer - Other', 'Printer - Other'),
+        ('Projector', 'Projector'),
+        ('Rack', 'Rack'),
+        ('Server - Blade', 'Server - Blade'),
+        ('Server - Rackmount', 'Server - Rackmount'),
+        ('Server - Tower', 'Server - Tower'),
+        ('Storage - Disk array', 'Storage - Disk array'),
+        ('Storage - NAS', 'Storage - NAS'),
+        ('Storage - SAN', 'Storage - SAN'),
+        ('Storage - Other', 'Storage - Other'),
+        ('Speaker', 'Speaker'),
+        ('Tablet', 'Tablet'),
+        ('Tape autoloader', 'Tape autoloader'),
+        ('Tape drive', 'Tape drive'),
+        ('UPS', 'UPS'),
+        ('Other', 'Other'),
     )
 
     model_type = models.CharField(max_length=50, choices=type_choices, verbose_name="type")
